pyproject.toml
[tool.poetry]
name = "django-drf"
version = "0.1.0"
description = ""
authors = ["Alexander Perov <perov97@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-filter = "^23.4"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
